Updated youtube-dl-exec, added deno js runtime
freyr-js has not been updated in a very long time. This is not intended as a comprehensive patch that will put it in an up-to-date, bleeding-edge state. Rather, this patch is intended to solve the key problems of using it to download songs from Spotify. Here are the following changes and why it was deliberately made that way:

- kept @ffmpeg/core *strictly* at version 0.11.0. Based from npm's repository, this is the only version of @ffmpeg/core that is in minor version 0.11. That means that the update after it is minor version 0.12 onwards, which has breaking changes.
- updated @ffmpeg/ffmpeg *strictly* at version 0.11.6. Same idea as the change above.
- updated youtube-dl-exec to v3.0.28: this is the underlying library used to interface with yt-dlp, and as such, needs to be updated.
- src/services/youtube.js updated to have the youtubedl object instantiation have the jsRuntime parameter (deno in this case). As for why it uses deno, it is the most recommended external JavaScript runtime for use with yt-dlp. The details for this is listed here:
https://github.com/yt-dlp/yt-dlp/issues/15012